# api/openapi.json
{
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      }
    }
  },
  "info": {
    "title": "Intelligent Data Pro API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
    "/api/v1/admin/audit/logs": {
      "get": {
        "description": "Expose the access-control audit trail (migrated from NexusCore).",
        "operationId": "list_audit_logs_api_v1_admin_audit_logs_get",
        "parameters": [
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 100,
              "title": "Limit",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List Audit Logs",
        "tags": [
          "admin"
        ]
      }
    },
    "/api/v1/admin/restart": {
      "post": {
        "description": "Restart systemd unit for 'web' or 'bot' (requires sudoers setup).\n\nReturns JSON with result. The service user must have passwordless sudo for\nthe corresponding units, e.g. in /etc/sudoers.d/intData:\n    www-data ALL=NOPASSWD: /bin/systemctl restart intdata-web, /bin/systemctl restart intdata-bot",
        "operationId": "restart_service_api_v1_admin_restart_post",
        "parameters": [
          {
            "in": "query",
            "name": "target",
            "required": true,
            "schema": {
              "title": "Target",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Restart Service",
        "tags": [
          "admin"
        ]
      }
    },
    "/api/v1/admin/role/{telegram_id}": {
      "post": {
        "operationId": "api_change_user_role_api_v1_admin_role__telegram_id__post",
        "parameters": [
          {
            "in": "path",
            "name": "telegram_id",
            "required": true,
            "schema": {
              "title": "Telegram Id",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "role",
            "required": true,
            "schema": {
              "title": "Role",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Api Change User Role",
        "tags": [
          "admin"
        ]
      }
    },
...
    "/api/v1/admin/web/link": {
      "post": {
        "operationId": "api_link_web_user_api_v1_admin_web_link_post",
        "parameters": [
          {
            "in": "query",
            "name": "web_user_id",
            "required": true,
            "schema": {
              "title": "Web User Id",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "tg_user_id",
            "required": true,
            "schema": {
              "title": "Tg User Id",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Api Link Web User",
        "tags": [
          "admin"
        ]
      }
    },
    "/api/v1/admin/web/role/{user_id}": {
      "post": {
        "operationId": "api_change_web_user_role_api_v1_admin_web_role__user_id__post",
        "parameters": [
          {
            "in": "path",
            "name": "user_id",
            "required": true,
            "schema": {
              "title": "User Id",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "role",
            "required": true,
            "schema": {
              "title": "Role",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Api Change Web User Role",
        "tags": [
          "admin"
        ]
      }
    },
    "/api/v1/admin/web/unlink": {
      "post": {
        "operationId": "api_unlink_web_user_api_v1_admin_web_unlink_post",
        "parameters": [
          {
            "in": "query",
            "name": "web_user_id",
            "required": true,
            "schema": {
              "title": "Web User Id",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "tg_user_id",
            "required": true,
            "schema": {
              "title": "Tg User Id",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Api Unlink Web User",
        "tags": [
          "admin"
        ]
      }
    },
...
  },
  "servers": [
    {
      "url": "/api/v1"
    }
  ],
...
}